+++ cfme/tests/services/test_service_catalogs.py
@@ -9,8 +9,10 @@ from cfme.infrastructure.provider.rhevm import RHEVMProvider
 from cfme.services.catalogs.catalog_item import CatalogBundle, CatalogItem, EditCatalogItemView
 from cfme.services.service_catalogs import ServiceCatalogs
 from cfme.utils import error
+from cfme.utils.appliance.implementations.ui import navigate_to
 from cfme.utils.log import logger
 from cfme.utils.wait import wait_for_decorator
+
 
 pytestmark = [
     pytest.mark.meta(server_roles="+automate"),
@@ -156,3 +158,17 @@ def test_request_with_orphaned_template(appliance, provider, setup_provider, cat
     provider.wait_for_delete()
     provision_request.wait_for_request(method='ui')
     assert provision_request.row.status.text == 'Error'
+
+
+@pytest.mark.tier(3)
+def test_advanced_search_registry_element(appliance, request):
+    """
+        Go to Services -> Workloads
+        Advanced Search -> Registry element
+        Element types select bar shouldn't disappear.
+    """
+    view = navigate_to(appliance.server, 'WorkloadsDefault')
+    view.search.open_advanced_search()
+    request.addfinalizer(view.search.close_advanced_search)
+    view.search.advanced_search_form.search_exp_editor.registry_form_view.fill({'type': "Registry"})
+    assert view.search.advanced_search_form.search_exp_editor.registry_form_view.type.is_displayed
